[FIX]: be more strict on anonymous links
Anonymous links shall not be prefixed by anything. If a word ends
with __ (double underscore) no other chars should be prepending
other than a white space or a line start.

The links are now part of the document context as they should be reset
after every document. This removes the need for extra resets in between.
