+++ packages/guides-restructured-text/src/RestructuredText/Parser/Productions/LinkRule.php
@@ -53,7 +53,7 @@ final class LinkRule implements Rule
         }
 
         //TODO: pass link object to setLink
-        $blockContext->getDocumentParserContext()->getContext()->setLink($link->getName(), $link->getUrl());
+        $blockContext->getDocumentParserContext()->setLink($link->getName(), $link->getUrl());
 
         return $node;
     }
