+++ tests/Functional/tests/anonymous/anonymous.html
@@ -1,2 +1,1 @@
-SKIP anonymous references don't work
-<p>I love <a href="https://github.com/">GitHub</a>, <a href="https://gitlab.com/">GitLab</a> and <a href="https://facebook.com/">Facebook</a>. But I don't affect references to __CLASS__.</p>
+<p>I love <a href="https://github.com/">GitHub</a>, <a href="https://gitlab.com/">GitLab</a> and <a href="https://facebook.com/">Facebook</a>. But I don&#039;t affect references to __CLASS__.</p>
