+++ tests/Functional/tests/anonymous/anonymous.rst
@@ -1,5 +1,5 @@
 I love GitHub__, GitLab__ and `Facebook`__. But I don't affect references to __CLASS__.
 
-.. __: https://github.com/
+__ https://github.com/
 .. __: https://gitlab.com/
 .. __: https://facebook.com/
