--- tests/Functional/tests/image-inline-anonymous/image-inline-anonymous.html
@@ -1,2 +1,0 @@
-SKIP anonymous inline images do not work
-<p>This is also an<img src="test.jpg"/>image.</p>
