--- tests/Functional/tests/image-inline-anonymous/image-inline-anonymous.rst
@@ -1,4 +1,0 @@
-SKIP anonymous inline images do not work
-This is also an :variable:`inline` image.
-
-.. |inline| image:: test.jpg
